fix: rebuild db and hide sold old product, group products

+++ sources/controllers/webhook_controller.js
@@ -1,5 +1,5 @@
 const AppString = require("../common/app_string");
-const { writeFile, baseRespond } = require("../common/functions");
+const { writeFile, baseRespond, mongoProductFromKiotVietProduct, lowercaseKey } = require("../common/functions");
 const mongoProduct = require("../models/mongo/mongo.product");
 const mongoCategory = require("../models/mongo/mongo.category");
 const KiotvietAPI = require("../common/kiotviet_api");
@@ -10,6 +10,7 @@ class WebhookController {
     async updateProduct(req, res, next) {
         try {
             let body = req.body;
+            console.log(body);
             let newProduct;
             try {
                 newProduct = body?.Notifications[0]?.Data[0];
@@ -17,15 +18,33 @@ class WebhookController {
                 newProduct == null;
             }
             if (newProduct != null) {
-                let updateFields = {
-                    skuCode: newProduct.Code,
-                    name: newProduct.Name,
-                    fullName: newProduct.FullName,
-                    categoryId: newProduct.CategoryId
+                newProduct = lowercaseKey(newProduct)
+                let checkProduct = await mongoProduct.findOne({ _id: newProduct.Id })
+                if (checkProduct != null) {
+                    let updateFields = {
+                        masterProductId : newProduct.MasterProductId,
+                        skuCode: newProduct.Code,
+                        name: newProduct.Name,
+                        fullName: newProduct.FullName,
+                        categoryId: newProduct.CategoryId,
+                        onHand : mongoProductFromKiotVietProduct(newProduct).onHand
+                    }
+                    await mongoProduct.findOneAndUpdate({
+                        _id: newProduct.Id
+                    }, updateFields)
+                } else {
+                    newProduct = lowercaseKey(newProduct)
+                    let mongoProduct = mongoProductFromKiotVietProduct(newProduct);
+                    await mongoProduct.create(newProduct)
                 }
-                await mongoProduct.findOneAndUpdate({
-                    _id: newProduct.Id
-                }, updateFields)
+                // update total onHand
+                let mongoNewProd = mongoProductFromKiotVietProduct(newProduct)
+                let masterId = newProduct.MasterProductId != null ? newProduct.MasterProductId :newProduct.Id;
+                let oldTotal = checkProduct?.onHand?? 0
+                let newTotal = mongoNewProd.onHand
+                WebhookController.updateTotalOnHand(masterId,oldTotal,newTotal)
+                // end of update total onHand
+
                 let category = await mongoCategory.findById(newProduct.CategoryId)
                 if (category == null) {
                     let response = await KiotvietAPI.callApi(ApiUrl.getCategoryById(newProduct.CategoryId))
@@ -35,7 +54,7 @@ class WebhookController {
                         parentId: 0
                     })
                 }
-                // await writeFile('./sources/public/update-product.json', JSON.stringify(body))
+                await writeFile('./sources/public/update-product.json', JSON.stringify(body))
                 res.json(baseRespond(true, AppString.ok))
             } else {
                 throw 'Incorrect format'
@@ -49,6 +68,7 @@ class WebhookController {
     async deleteProduct(req, res, next) {
         try {
             let body = req.body;
+            console.log(body)
             let deletedId;
             try {
                 deletedId = body?.RemoveId;
@@ -56,12 +76,16 @@ class WebhookController {
                 deletedId = null
             }
             if (deletedId != null) {
-                await mongoProduct.deleteOne({
-                    _id: deletedId
-                })
-                await mongoProductAttribute.deleteMany({
-                    productId: deletedId
-                })
+                let checkProduct = await mongoProduct.findOne({ _id: deletedId })
+                if (checkProduct != null) {
+                    await mongoProduct.deleteOne({
+                        _id: deletedId
+                    })
+                    await mongoProductAttribute.deleteMany({
+                        productId: deletedId
+                    })
+                }
+
                 // await writeFile('./sources/public/delete-product.json', JSON.stringify(body))
                 res.json(baseRespond(true, AppString.ok))
 
@@ -73,6 +97,66 @@ class WebhookController {
             res.status(400)
             res.json(baseRespond(false, err))
         }
+    }
+    async updateOnHand(req, res, next) {
+        try {
+            let body = req.body;
+            console.log(body);
+            let newProduct;
+            try {
+                newProduct = body?.Notifications[0]?.Data[0];
+            } catch (err) {
+                newProduct == null;
+            }
+            if (newProduct != null) {
+                let updateFields = {
+                    onHand: newProduct.OnHand,
+                }
+                let checkProduct = await mongoProduct.findOne({ _id: newProduct.ProductId })
+                if (checkProduct != null) {
+                    await mongoProduct.findOneAndUpdate({
+                        _id: newProduct.ProductId
+                    }, updateFields)
+
+                // update total onHand
+                let masterId = checkProduct.masterProductId != null ? checkProduct.masterProductId :checkProduct._id;
+                let oldTotal = checkProduct?.onHand?? 0
+                let newTotal = newProduct.OnHand
+                WebhookController.updateTotalOnHand(masterId,oldTotal,newTotal)
+                // end of update total onHand
+                }
+                
+                await writeFile('./sources/public/update-onhand.json', JSON.stringify(body))
+                res.json(baseRespond(true, AppString.ok))
+            } else {
+                throw 'Incorrect format'
+            }
+        } catch (err) {
+            console.error(err);
+            res.status(400)
+            res.json(baseRespond(false, err))
+        }
+    }
+    static async updateTotalOnHand(masterId,oldTotal,newTotal) {
+        if (oldTotal < 0) oldTotal = 0
+        if (newTotal < 0) newTotal = 0
+        let changeAmmount = newTotal - oldTotal;
+        if (changeAmmount == 0) return;
+        let masterProduct = await mongoProduct.findOne({ _id: masterId })
+        if (masterProduct != null) {
+            let totalOnHand = masterProduct.totalOnHand ?? 0;
+            totalOnHand += changeAmmount;
+            if (totalOnHand >= 0) {
+                await mongoProduct.updateOne({ _id: masterId }, {
+                     $set: { totalOnHand: totalOnHand}
+                })
+            }
+
+        }
+        
+
+
+        
     }
 }
